Добавил модульность и поправил функции
