+++ js/utils.js
@@ -1,0 +1,41 @@
+const getRandomInteger = (min, max) => {
+  const lower = Math.min(Math.abs(min), Math.abs(max));
+  const upper = Math.max(Math.abs(min), Math.abs(max));
+  return Math.floor(Math.random() * (upper - lower + 1)) + lower;
+};
+
+const getRandomFloat = (min, max, precision) => {
+  const lower = Math.min(Math.abs(min), Math.abs(max));
+  const upper = Math.max(Math.abs(min), Math.abs(max));
+  return parseFloat((Math.random() * (upper - lower) + lower).toFixed(precision));
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0,elements.length - 1)];
+const getRandomArrayQuantity = (elements) => getRandomInteger(1,elements.length);
+
+const getRandomArrayItems = (items,itemQuantity,deleteFromOriginal) => {
+  if (items.length === 0) {
+    throw new Error('Массив пустой!');
+  }
+  itemQuantity=Math.abs(itemQuantity);
+  const itemsRemaining = items.slice();
+  const deleteFromArray = () => {
+    let arrayToDeleteFrom;
+    if (deleteFromOriginal) {
+      arrayToDeleteFrom = items;
+    } else {
+      arrayToDeleteFrom = itemsRemaining;
+    }
+    const itemSeed = getRandomInteger(0,arrayToDeleteFrom.length-1);
+    const currentItem = arrayToDeleteFrom[itemSeed];
+    arrayToDeleteFrom.splice(itemSeed,1);
+    return currentItem;
+  };
+  if (itemQuantity === 1) {
+    return deleteFromArray();
+  }
+  const data = new Array(itemQuantity).fill(null).map(deleteFromArray);
+  return data;
+};
+
+export {getRandomInteger, getRandomFloat, getRandomArrayElement, getRandomArrayQuantity, getRandomArrayItems};
